requirements.txt
ultralytics
opencv-python
PyYAML
matplotlib
Pillow
ffmpeg-python
tqdm
numpy
pandas
torch
torchinfo
torchvision